fix: allow Settings External MCP approvals and move session sync to application state
Accept approval responses from Settings, clear sibling approval cards on resolve, register Always Allow grant persisters outside Catty, and keep session sync behind the application/state lint boundary.
